Homework-5. Adding the ability to share events

+++ controllers/CalendarController.php
@@ -2,10 +2,13 @@
 
 namespace app\controllers;
 
+use app\models\Access;
 use Yii;
 use app\models\Calendar;
 use app\models\search\SearchCalendar;
+use yii\filters\AccessControl;
 use yii\web\Controller;
+use yii\web\ForbiddenHttpException;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 use yii2fullcalendar\models\Event;
@@ -21,6 +24,17 @@ class CalendarController extends Controller
     public function behaviors()
     {
         return [
+            'access' => [
+                'class' => AccessControl::className(),
+                'only' => ['mynotes', 'create', 'update', 'delete'],
+                'rules' => [
+                    [
+                        'actions' => ['mynotes', 'create', 'update', 'delete'],
+                        'allow' => true,
+                        'roles' => ['@'],
+                    ],
+                ],
+            ],
             'verbs' => [
                 'class' => VerbFilter::className(),
                 'actions' => [
@@ -36,12 +50,21 @@ class CalendarController extends Controller
      */
     public function actionIndex()
     {
+        if (\Yii::$app->user->isGuest) {
+            return $this->redirect(['viewguest']);
+        }
         $searchModel = new SearchCalendar();
-        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
-        $e = Calendar::find()->all();
+        $dataProvider = $searchModel->search(
+            [
+                'SearchCalendar' =>
+                  ['creator' => Yii::$app->user->id]
+
+            ]
+        );
+        $e = Calendar::find()->where(['creator' => Yii::$app->user->id])->all();
         foreach ($e as $eve)
         {
-            $event = new \yii2fullcalendar\models\Event();
+            $event = new Event();
             $event->id = $eve->id;
             $event->title = $eve->text;
             $event->start = $eve->date_event;
@@ -50,7 +73,7 @@ class CalendarController extends Controller
         }
 
         if(!$e){
-            $Event = new \yii2fullcalendar\models\Event();
+            $Event = new Event();
             $Event->id = 1;
             $Event->title = 'Сегодня';
             $Event->start = date('Y-m-d\TH:i:s\Z');
@@ -58,11 +81,59 @@ class CalendarController extends Controller
         }
         return $this->render('index', [
             'searchModel' => $searchModel,
+            'creator' => Yii::$app->user->id,
             'dataProvider' => $dataProvider,
             'events' => $events
         ]);
     }
 
+    public function actionViewguest(){
+        $searchModel = new SearchCalendar();
+        $dataProvider = $searchModel->search([
+
+        ]);
+        $Event = new Event();
+        $Event->id = 1;
+        $Event->title = 'Сегодня';
+        $Event->start = date('Y-m-d\TH:i:s\Z');
+        $events[] = $Event;
+        return $this->render('index', [
+            'searchModel' => $searchModel,
+            'dataProvider' => $dataProvider,
+            'events' => $events
+        ]);
+    }
+
+    public function actionFriendcalendars($id){
+        $searchModel = new SearchCalendar();
+        $dataProvider = $searchModel->search([
+            'SearchCalendar' => [
+                'creator' => $id,
+                'access' => [
+                    'user_guest' => Yii::$app->user->id,
+                ]
+            ]
+        ]);
+        $e = Calendar::find()->select('*')
+                        ->leftJoin('table_access', '`table_access`.`date` = `clndr_calendar`.`date_event`')
+                        ->where(['user_guest'=>Yii::$app->user->id])
+                        ->all();
+        foreach ($e as $eve)
+        {
+            $event = new Event();
+            $event->id = $eve->id;
+            $event->title = $eve->text;
+            $event->start = $eve->date_event;
+            $event->backgroundColor = 'red';
+            $events[] = $event;
+        }
+
+        return $this->render('index', [
+            'searchModel' => $searchModel,
+            'dataProvider' => $dataProvider,
+            'events' => $events
+        ]);
+    }
     /**
      * Displays a single Calendar model.
      * @param integer $id
@@ -70,9 +141,23 @@ class CalendarController extends Controller
      */
     public function actionView($id)
     {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
+        $model = $this->findModel($id);
+        $result = Access::checkAccess($model);
+        switch($result){
+            case Access::ACCESS_CREATOR:
+                return $this->render('viewCreator', [
+                    'model' => $model,
+                ]);
+                break;
+            case Access::ACCESS_GUEST:
+                return $this->render('viewGuest', [
+                    'model' => $model,
+                ]);
+                break;
+            default:
+                throw new ForbiddenHttpException("Access denied", 403);
+                break;
+        }
     }
 
     /**
@@ -101,6 +186,9 @@ class CalendarController extends Controller
      */
     public function actionUpdate($id)
     {
+        if (\Yii::$app->user->isGuest) {
+            return $this->redirect(['index']);
+        }
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -120,6 +208,9 @@ class CalendarController extends Controller
      */
     public function actionDelete($id)
     {
+        if (\Yii::$app->user->isGuest) {
+            return $this->redirect(['index']);
+        }
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
